pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spriter"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Dane Howard", email = "mirrord@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
]

[project.urls]
Documentation = "https://github.com/Dane Howard/spriter#readme"
Issues = "https://github.com/Dane Howard/spriter/issues"
Source = "https://github.com/Dane Howard/spriter"

[tool.hatch.version]
path = "src/spriter/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/spriter tests}"

[tool.coverage.run]
source_pkgs = ["spriter", "tests"]
branch = true
parallel = true
omit = [
  "src/spriter/__about__.py",
]

[tool.coverage.paths]
spriter = ["src/spriter", "*/spriter/src/spriter"]
tests = ["tests", "*/spriter/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
